backend/requirements.txt
fastapi
uvicorn
python-multipart
transformers
torch
torchvision
pillow
timm
